+++ src/Symfony/Bundle/FrameworkBundle/Tests/Templating/Helper/FormHelperTableLayoutTest.php
@@ -16,9 +16,11 @@ use Symfony\Bundle\FrameworkBundle\Templating\Helper\TranslatorHelper;
 use Symfony\Bundle\FrameworkBundle\Tests\Templating\Helper\Fixtures\StubTemplateNameParser;
 use Symfony\Bundle\FrameworkBundle\Tests\Templating\Helper\Fixtures\StubTranslator;
 use Symfony\Component\Form\FormView;
+use Symfony\Component\Form\FormRenderer;
+use Symfony\Component\Form\Extension\Templating\TemplatingRendererEngine;
+use Symfony\Component\Form\Tests\AbstractTableLayoutTest;
 use Symfony\Component\Templating\PhpEngine;
 use Symfony\Component\Templating\Loader\FilesystemLoader;
-use Symfony\Component\Form\Tests\AbstractTableLayoutTest;
 
 class FormHelperTableLayoutTest extends AbstractTableLayoutTest
 {
@@ -34,11 +36,13 @@ class FormHelperTableLayoutTest extends AbstractTableLayoutTest
         $loader = new FilesystemLoader(array());
         $engine = new PhpEngine($templateNameParser, $loader);
         $engine->addGlobal('global', '');
-
-        $this->helper = new FormHelper($engine, $this->getMock('Symfony\Component\Form\Extension\Csrf\CsrfProvider\CsrfProviderInterface'), array(
+        $rendererEngine = new TemplatingRendererEngine($engine, array(
             'FrameworkBundle:Form',
             'FrameworkBundle:FormTable'
         ));
+        $renderer = new FormRenderer($rendererEngine, $this->getMock('Symfony\Component\Form\Extension\Csrf\CsrfProvider\CsrfProviderInterface'));
+
+        $this->helper = new FormHelper($renderer);
 
         $engine->setHelpers(array(
             $this->helper,